+++ src/main/java/roomescape/infrastructure/JdbcReservationRepositoryImpl.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.infrastructure;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
