+++ src/main/java/roomescape/infrastructure/JdbcThemeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.infrastructure;
 
 import java.util.List;
 import java.util.Map;
